authentication/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
decorator==5.1.1
future==0.18.2
geocoder==1.38.1
idna==3.3
importlib-metadata==4.11.1
oauthlib==3.1.1
pkg-resources==0.0.0
python-dotenv==0.19.2
ratelim==0.1.6
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==4.4.0
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
